moved Norm and NormStats utility methodes to rlwe/utils.go
