+++ rlwe/decryptor.go
@@ -1,9 +1,6 @@
 package rlwe
 
 import (
-	"math"
-	"math/big"
-
 	"github.com/tuneinsight/lattigo/v4/ring"
 	"github.com/tuneinsight/lattigo/v4/utils"
 )
@@ -111,78 +108,4 @@ func (d *decryptor) WithKey(sk *SecretKey) Decryptor {
 		buff:  d.ringQ.NewPoly(),
 		sk:    sk,
 	}
-}
-
-// Norm returns the log2 of the standard deviation, minimum and maximum absolute norm of
-// the decrypted ciphertext, before the decoding (i.e. including the error).
-func Norm(ct *Ciphertext, dec Decryptor) (std, min, max float64) {
-
-	params := dec.(*decryptor).params
-
-	coeffsBigint := make([]*big.Int, params.N())
-	for i := range coeffsBigint {
-		coeffsBigint[i] = new(big.Int)
-	}
-
-	pt := NewPlaintext(params, ct.Level())
-
-	dec.Decrypt(ct, pt)
-
-	if pt.IsNTT {
-		params.RingQ().InvNTTLvl(ct.Level(), pt.Value, pt.Value)
-	}
-
-	params.RingQ().PolyToBigintCenteredLvl(ct.Level(), pt.Value, 1, coeffsBigint)
-
-	return NormStats(coeffsBigint)
-}
-
-func NormStats(vec []*big.Int) (float64, float64, float64) {
-
-	vecfloat := make([]*big.Float, len(vec))
-	minErr := new(big.Float).SetFloat64(0)
-	maxErr := new(big.Float).SetFloat64(0)
-	tmp := new(big.Float)
-	minErr.SetInt(vec[0])
-	minErr.Abs(minErr)
-	for i := range vec {
-		vecfloat[i] = new(big.Float)
-		vecfloat[i].SetInt(vec[i])
-
-		tmp.Abs(vecfloat[i])
-
-		if minErr.Cmp(tmp) == 1 {
-			minErr.Set(tmp)
-		}
-
-		if maxErr.Cmp(tmp) == -1 {
-			maxErr.Set(tmp)
-		}
-	}
-
-	n := new(big.Float).SetFloat64(float64(len(vec)))
-
-	mean := new(big.Float).SetFloat64(0)
-
-	for _, c := range vecfloat {
-		mean.Add(mean, c)
-	}
-
-	mean.Quo(mean, n)
-
-	err := new(big.Float).SetFloat64(0)
-	for _, c := range vecfloat {
-		tmp.Sub(c, mean)
-		tmp.Mul(tmp, tmp)
-		err.Add(err, tmp)
-	}
-
-	err.Quo(err, n)
-	err.Sqrt(err)
-
-	x, _ := err.Float64()
-	y, _ := minErr.Float64()
-	z, _ := maxErr.Float64()
-
-	return math.Log2(x), math.Log2(y), math.Log2(z)
 }
